ggheatmap additional themes

These themes are provided as helpers to build panels with ggheatmap. They are all built on top of ggplot2::theme_minimal().

theme_sparse(...)

theme_sparse2(...)

theme_quant(...)

theme_scatter(...)

Arguments

...
all arguments passed to ggplot2::theme_minimal().

Details

theme_sparse() is the baseline simple and spartan theme to build plots that align with columns of ggheatmaps. theme_sparse2() is designed for plots that will align to rows of the ggheatmap. theme_quant() adds horizontal lines to theme_sparse() and can be a useful upgrade for displaying numeric values in the y-axis. theme_scatter() is a all-purpose theme for building ggplots.